pyproject.toml
[project]
name = "captcha-recognizer"
version = "0.1.0"
description = "CAPTCHA Recognition System"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "hydra-colorlog>=1.2.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.0.0",
    "dvc>=3.0.0",
    "dvc-gdrive>=2.0.0",
    "fire>=0.5.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "albumentations>=1.3.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "kagglehub>=0.1.0",
    "lion-pytorch>=0.1.0",
    "pathlib>=1.0.1",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "protobuf>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0.0",
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

export = [
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "protobuf>=3.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = "E203,W503"
exclude = [".git", "__pycache__", ".venv", "venv"]